requirements.txt
requests
tldextract
python-whois
dnspython
selenium
webdriver-manager
ndjson